requirements.txt
plexapi
requests
beautifulsoup4
PyYAML
apscheduler
tzdata
